Web application

+++ include/library/Game/Handler.php
@@ -2,11 +2,13 @@
 
 namespace Clever\Library\Game;
 
+use Mexenus\Database\Database;
+
 use Clever\Library\Config;
-use Clever\Library\Database;
 use Clever\Library\Encryption;
 
 use Workerman\Worker;
+use Workerman\Timer;
 use Workerman\Connection\TcpConnection;
 
 use Clever\Library\Model\Game;
@@ -22,7 +24,7 @@ class Handler
 
 
 	/**
-	 * Clever\Library\Database
+	 * Mexenus\Database\Database
 	 * 
 	 * @var Database
 	 */
@@ -70,6 +72,22 @@ class Handler
 
 
 	/**
+	 * Contains all the game instances.
+	 *  
+	 * @var array
+	 */
+	private $instances = [];
+
+
+	/**
+	 * ID of the the timer used to clear the unused objects.
+	 *  
+	 * @var int
+	 */
+	private $timerID = null;
+
+
+	/**
 	 * @param Clever\Library\Config $config
 	 * 
 	 * @return void
@@ -90,15 +108,20 @@ class Handler
 	 */
 	public function onWorkerStart()
 	{
+		$this->file = tmpfile();
+		$this->file2 = tmpfile();
 
-		$this->database = new Database($this->config);
-		$this->database->timeoutError = false;
+		$data = "#!/usr/bin/env php\n" . file_get_contents('/var/www/html/that-is-clever/test.php');
+
+		fwrite($this->file, $data);
+		fwrite($this->file2, $data);
+
+		$this->database = new Database($this->config->get('db_host'), $this->config->get('db_name'), $this->config->get('db_user'), $this->config->get('db_pass'));
 
 		$this->gameDB = new Game($this->database);
 		
 		$this->SoloHandler = new SoloHandler($this->database);
 		$this->FriendHandler = new FriendHandler($this->database);
-		
 	}
 
 
@@ -111,31 +134,47 @@ class Handler
 	 */
 	public function onWebSocketConnect(TcpConnection $connection)
 	{
+		if (!$this->timerID) $this->timerID = Timer::add(3600, [$this, 'clearUnusedObjects']);
+
+
+		if (isset($_GET['bot'])) return $this->incomingBotRequest($connection);
+
+
 		if (!isset($_GET['token'], $_COOKIE['serial'], $_COOKIE['game_id'])) {
 
 			$connection->destroy();
 			return;
 		}
 
-		$gameID = $this->crypto->decryptString($_COOKIE['game_id']);
 
-		$gameData = $this->gameDB->preGameDataByID($gameID);
+		try {
 
-		if (!$gameData) {
+			$gameID = $this->crypto->decryptString($_COOKIE['game_id']);
+
+		} catch (\Exception $e) {
+
+			$connection->destroy();
+			return;
+		}
+		
+
+		$game = $this->gameDB->find($gameID);
+
+		if (!$game || $game->status != 2) {
 			
 			$connection->destroy();
 			return;
 		}
 
-		switch ($gameData->type) {
+		switch ($game->type) {
 			case 3:
 				
-				$result = $this->SoloHandler->newConnection($connection, $gameData);
+				$object = $this->SoloHandler->newConnection($connection, $game);
 
 			break;
 			case 2:
 
-				$result = $this->FriendHandler->newConnection($connection, $gameData);
+				$object = $this->FriendHandler->newConnection($connection, $game);
 				
 			break;
 			default:
@@ -145,27 +184,40 @@ class Handler
 
 
 		}
+
+		$this->instances[] = $object;
 	}
 
 
-	/**
-	 * Function on message.
-	 * 
-	 * @return void
-	 */
-	public function onMessage($connection, $data)
-	{
-
-	}
-
-
-	/**
-	 * Function on connection closing.
-	 * 
-	 * @return void
-	 */
-	public function onClose($connection)
+	private function incomingBotRequest(TcpConnection $connection)
 	{
 		
+	}
+
+
+	private function sleep()
+	{
+		Timer::del($this->timerID);
+		$this->timerID = null;
+
+		$this->instances = [];
+
+		$this->database->sleep();
+	}
+
+
+	/**
+	 * Garbage collector. Clear finished games.
+	 * 
+	 * @return void
+	 */
+	public function clearUnusedObjects()
+	{
+		if (empty($this->instances)) return $this->sleep();
+
+		foreach ($this->instances as $key => $value) {
+			
+			if ($this->instances[$key]->destroy) unset($this->instances[$key]);
+		}
 	}
 }
